refactor: event dto 위치 변경
