+++ order/src/main/java/store/_0982/order/application/order/OrderPointListener.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
-import store._0982.common.event.dto.OrderCreatedEvent;
+import store._0982.order.event.dto.OrderCreatedEvent;
 import store._0982.order.infrastructure.client.payment.PaymentClient;
 import store._0982.order.infrastructure.client.payment.dto.PointDeductRequest;
